Cput

+++ src/gifti/gifti_test.h
@@ -15,7 +15,7 @@
       } while(0)
 
 int ewrite_data_line (void *, int, int, int, int, int, FILE *);
-int ewrite_many_lines(void **, int, size_t, size_t, int, FILE *);
+int ewrite_many_lines(void **, int, long long, long long, int, FILE *);
 int write_1D_file    (giiDataArray **, int, char *, int);
 int write_as_ascii   (gifti_image *, char *);
 int write_surf_file  (giiDataArray *, giiDataArray *, char *, int);
